added email confirmation on signup and updated purchase confirmation email with reservation info
