+++ server/transactions.js
@@ -5,22 +5,10 @@ const Sequelize = require('sequelize')
 const Home = db.model('homes')
 const Transaction = db.model('transactions')
 const User = db.model('users')
+const moment = require('moment')
 const nodemailer = require('nodemailer');
+const transporter = require('./transporter')
 const {mustBeLoggedIn, forbidden} = require('./auth.filters')
-
-const transporter = nodemailer.createTransport({
-    service: 'gmail',
-    auth: {
-      type: 'OAuth2',
-        // xoauth2: xoauth2.createXOAuth2Generator({
-            user: 'colickyboy@gmail.com',
-            clientId: '540263018909-svsmsmsj0vjmmctjgd7o0m9pliu0g9v3.apps.googleusercontent.com',
-            clientSecret: 'nsh8SVcFYKwy4kKflIZgQrCf',
-            refreshToken: '1/V3kw9UjXbVImh5qQV3w43_mAZZ8XoVbg4uF63bg51vk',
-            accessToken: 'ya29.GltoBIuYWqOyq-FIaQfIZgs2xSH92g0SVfAzrIceh1z4zJxK0NMFBg3QlGxZn64v25WUfrPq_yQfjrLpewPE3Ujk6WiOKiTFJ1cCMA06qnnwBBHv6IbQGvLTbQtU'
-        // })
-    }
-})
 
 module.exports = require('express').Router({mergeParams: true})
   .get('/guestfuture', (req, res, next) => {
@@ -32,7 +20,6 @@ module.exports = require('express').Router({mergeParams: true})
       include: [{all: true}]
     })
     .then(futureGuestTransactions => {
-      console.log("futureGuestTransactions", futureGuestTransactions)
       res.json(futureGuestTransactions)
     })
     .catch(next)
@@ -46,7 +33,6 @@ module.exports = require('express').Router({mergeParams: true})
       include: [{all: true}]
     })
     .then(pastGuestTransactions => {
-      console.log("pastGuestTransactions", pastGuestTransactions)
       res.json(pastGuestTransactions)
     })
     .catch(next)
@@ -79,15 +65,11 @@ module.exports = require('express').Router({mergeParams: true})
     })
   .post('/',
     (req, res, next) => {
-      console.log('req.body',req.body)
-
       Transaction.create(req.body)
-      .then(res => {
-        console.log("res", res)
-        console.log("req.body", req.body)
+      .then(transactions => {
         const gettingHome = Home.findById(req.body.home_id)
+        const gettingHost = User.findById(req.body.host_id)
         const gettingUser = User.findById(req.body.guest_id)
-
         User.findById(req.body.guest_id)
         .then((user)=>{
           user.getCart()
@@ -95,18 +77,15 @@ module.exports = require('express').Router({mergeParams: true})
             cart.setAvailabilities([])
           })
         })
-
-
-        return Promise.all([res.data, gettingHome, gettingUser])
+        return Promise.all([transactions, gettingHome, gettingHost, gettingUser])
       })
-      .then(([transaction, home, user]) => {
-        console.log('stuff', transaction, home, user)
+      .then(([transaction, home, host, user]) => {
         res.status(201).json(transaction)
         var mailOptions = {
           from: 'GalacticBnB <colickyboy@gmail.com>',
-          to: 'robtong21@gmail.com',
+          to: user.email,
           subject: 'Your GalacticBnB reservation is set',
-          html: 'Thank you for reserving a home through GalacticBnB.<br/>We hope you enjoy your stay!'
+          html: '<p>Hello, ' + user.name + '!</p><p>Thank you for reserving a home through GalacticBnB.</p><p>This is your order confirmation. Please save this for future reference.</p><p>Reservation details:</p><p><strong>Staying at:</strong> ' + home.name + '</p><p><strong>Host:</strong> ' + host.name + '</p><p><strong>Price:</strong> ' + transaction.price + '</p><p><strong>Check-in:</strong> ' + moment(res.startDate, "MM/DD/YYYY") + '</p><p><strong>Checkout:</strong> ' + moment(res.endDate, "MM/DD/YYYY") + '</p><p>We hope you enjoy your stay at ' + home.name + '!</p><p>Sincerely,</p><p>The GalacticBnB Team</p>'
         }
         transporter.sendMail(mailOptions, function (err, res) {
             if(err){
